### 2025-02-20 #7817 @editorialbot
[PRE REVIEW]: PAAT: Physical Activity Analysis Toolbox for the analysis of hip-worn raw accelerometer data in Python
**Submitting author:** @trybnetic (<a href="http://orcid.org/0009-0006-8225-833X">Marc Weitz</a>)
**Repository:** https://github.com/Trybnetic/paat
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.0.0b8
