+++ paper/paper.bib
@@ -206,15 +206,17 @@ Section: Journal for the Measurement of Physical Behaviour},
 	file = {Hildebrand et al. - 2014 - Age Group Comparability of Raw Accelerometer Outpu.pdf:/home/msw/Zotero/storage/3XC4VTU3/Hildebrand et al. - 2014 - Age Group Comparability of Raw Accelerometer Outpu.pdf:application/pdf;Snapshot:/home/msw/Zotero/storage/R4X58EKW/Age_Group_Comparability_of_Raw_Accelerometer.17.html:text/html},
 }
 
-@misc{weitz_automatic_2024,
-	title = {{Automatic} time in bed detection from hip-worn accelerometers for large epidemiological studies: {The} {Tromsø} {Study}},
-	abstract = {Accelerometers are frequently used to assess physical activity in large epidemiological studies. They can monitor movement patterns and cycles over several days under free-living conditions and are usually either worn on the wrist or the hip. While wrist-worn accelerometers have been frequently used to additionally assess sleep and time in bed behavior, hip-worn accelerometers have been widely neglected for this task due to their primary focus on physical activity. Here, we present a new method to identify the time in bed to enable further analysis options for large-scale studies using hip-placement. We introduced new and accelerometer specific data augmentation methods to improve training and generalization performance. Subsequently, we trained a neural network model on a sample from the population-based Tromsø Study and evaluated it on two additional data sets. Our algorithm achieved an accuracy of 95\% on the training data, 92% on unseen data from the same population and comparable results to consumer-wearable data obtained from a demographically different population. Generalization performance was overall good, however, we found that on some particular days or participants, the trained model significantly over- or underestimated time in bed. Despite these limitations, we anticipate our approach to be a starting point for more sophisticated methods to identify time in bed or at some point even sleep from hip-worn acceleration signals. This can enable the re-use of already collected data, for example, for longitudinal analyses where sleep-related research questions only recently got into focus or sedentary time needs to be estimated in 24h wear protocols. },
-	language = {en},
-	journal = {PLOS ONE},
-	author = {Weitz, Marc and Syed, Shaheen and Hopstock, Laila A. and Morseth, Bente and Henriksen, André and Horsch, Alexander},
-	year = "under review",
+@article{weitz_automatic_2024,
+	title = {Automatic time in bed detection from hip-worn accelerometers for large epidemiological studies: {The} {Tromso} {Study}},
+	url = {http://medrxiv.org/content/early/2025/02/19/2022.03.07.22270992.abstract},
+	doi = {10.1101/2022.03.07.22270992},
[…]
