### 2025-02-20 #7817 @Trybnetic
Thank you for moving our submission further.

I have checked the list of potential reviewers and could not find anyone with "physical activity" or "accelerometer" as topic area. Despite the difference in research focus, I can imagine that the people stating "gait" or "gait analysis" might also be familiar with accelerometers and might thus be a good fit as reviewers.
