+++ paper/paper.bib
@@ -129,14 +129,14 @@ Section: Journal for the Measurement of Physical Behaviour},
 }
 
 @misc{chan_actipy_2024,
-	title = {Actipy: {Python} {Software} {Development} {Kit} ({SDK}) to process {Axivity}, {GENEActiv}, and {Actigraph} files},
-	shorttitle = {{OxWearables}/actipy},
-	url = {https://github.com/OxWearables/actipy},
-	urldate = {2024-05-01},
-	author = {Chan, Shing and Yuan, Hang and Doherty, Aiden},
-	month = may,
+	title = {actipy: {Python} toolkit to process wearable sensor data},
+	shorttitle = {actipy},
+	publisher = {Zenodo},
+	author = {Chan, Shing and Doherty, Aiden},
+	month = dec,
 	year = {2024},
-	keywords = {paat-paper},
+	doi = {10.5281/zenodo.14514488},
+	file = {Snapshot:/home/msw/Zotero/storage/HXYIKBTP/14514694.html:text/html},
 }
 
 
