+++ paper/paper.bib
@@ -1,4 +1,3 @@
-
 @article{syed_evaluating_2020,
 	title = {Evaluating the performance of raw and epoch non-wear algorithms using multiple accelerometers and electrocardiogram recordings},
 	volume = {10},
@@ -139,73 +138,6 @@ Section: Journal for the Measurement of Physical Behaviour},
 	file = {Snapshot:/home/msw/Zotero/storage/HXYIKBTP/14514694.html:text/html},
 }
 
-
-@article{hees_separating_2013,
-	title = {Separating {Movement} and {Gravity} {Components} in an {Acceleration} {Signal} and {Implications} for the {Assessment} of {Human} {Daily} {Physical} {Activity}},
-	volume = {8},
-	issn = {1932-6203},
-	url = {https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0061691},
-	doi = {10.1371/journal.pone.0061691},
-	abstract = {Introduction Human body acceleration is often used as an indicator of daily physical activity in epidemiological research. Raw acceleration signals contain three basic components: movement, gravity, and noise. Separation of these becomes increasingly difficult during rotational movements. We aimed to evaluate five different methods (metrics) of processing acceleration signals on their ability to remove the gravitational component of acceleration during standardised mechanical movements and the implications for human daily physical activity assessment. Methods An industrial robot rotated accelerometers in the vertical plane. Radius, frequency, and angular range of motion were systematically varied. Three metrics (Euclidian norm minus one [ENMO], Euclidian norm of the high-pass filtered signals [HFEN], and HFEN plus Euclidean norm of low-pass filtered signals minus 1 g [HFEN+]) were derived for each experimental condition and compared against the reference acceleration (forward kinematics) of the robot arm. We then compared metrics derived from human acceleration signals from the wrist and hip in 97 adults (22–65 yr), and wrist in 63 women (20–35 yr) in whom daily activity-related energy expenditure (PAEE) was available. Results In the robot experiment, HFEN+ had lowest error during (vertical plane) rotations at an oscillating frequency higher than the filter cut-off frequency while for lower frequencies ENMO performed better. In the human experiments, metrics HFEN and ENMO on hip were most discrepant (within- and between-individual explained variance of 0.90 and 0.46, respectively). ENMO, HFEN and HFEN+ explained 34\%, 30\% and 36\% of the variance in daily PAEE, respectively, compared to 26\% for a metric which did not attempt to remove the gravitational component (metric EN). Conclusion In conclusion, none of the metrics as evaluated systematically outperformed all other metrics across a wide range of standardised kinematic conditions. However, choice of metric explains different degrees of variance in daily human physical activity.},
-	language = {en},
-	number = {4},
-	urldate = {2024-06-28},
-	journal = {PLOS ONE},
-	author = {Van Hees, Vincent T. and Gorzelniak, Lukas and León, Emmanuel Carlos Dean and Eder, Martin and Pias, Marcelo and Taherian, Salman and Ekelund, Ulf and Renström, Frida and Franks, Paul W. and Horsch, Alexander and Brage, Søren},
-	month = apr,
-	year = {2013},
-	note = {Publisher: Public Library of Science},
-	keywords = {Acceleration, Accelerometers, Hip, paat, paat-paper, paper2, paper4, Physical activity, Robots, Signal filtering, Skeletal joints, Wrist},
-	pages = {e61691},
-	file = {Full Text PDF:/home/msw/Zotero/storage/CWJ6YCE2/Hees et al. - 2013 - Separating Movement and Gravity Components in an A.pdf:application/pdf},
-}
-
-
-@article{hildebrand_evaluation_2017,
[…]
