### 2025-04-06 #7817 @sappelhoff
Hi @Kevin-Mattheus-Moerman I would be willing to edit this, but given my most recent experiences with editing a submission, I would first like to receive an answer from @Trybnetic for the following questions.

1. The JOSS ["similarity suggestions"](https://github.com/openjournals/joss-reviews/issues/7817#issuecomment-2729974868) show two more papers that focus on accelerometer data analysis on the lower back (Sit2StandPy and GaitPy). This is arguably something else than "hip-worn" as for PAAT, but it seems to me that there are many accelerometer analysis packages that focus on specific use cases. What are your thoughts on this, and are you aware of any over-arching analysis packages that generalize across the specific use cases? Or do you believe that it is important to have separate analysis packages for each use case, as they are indeed so specific that it would not be sensible to generalize?

2. Related to the first question, how do you judge the likelihood of other researchers (re-)using PAAT, rather than creating their own software suite? Have you had colleagues from outside your own lab reach out to you and try to use your software?

3. What are your future plans for the software (1 year, 5 years, ...)?
