### 2025-04-21 #7817 @sappelhoff
Thank you @Trybnetic, I am happy to take this on as an editor.

Do you have suggestions on who may be a suitable peer-reviewer? (taking into account our [COI policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy))
