### 2025-04-23 #7817 @Trybnetic
> Do you have suggestions on who may be a suitable peer-reviewer? (taking into account our [COI policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy))

I browsed a little bit the [JOSS reviewer database](https://reviewers.joss.theoj.org/reviewers/) and while there was no one who stated physical activity as a topic area, I found some potential reviewers that might have experiences with accelerometers in other applications and, thus, might be more familiar with the topic. E.g., rmndrs89, akuederle, mprib, marcoghislieri

Further keywords that might provide suitable reviewers might be "gait", "Epidemiology", or "Public Health".

Please let me know if you need more suggestions.
