### 2025-04-24 9b0d004
adds python 3.12
