+++ pyproject.toml
@@ -27,7 +27,7 @@ classifiers = [
 ]
 
 [tool.poetry.dependencies]
-python = ">3.9,<3.12"  # Compatible python versions must be declared here
+python = ">3.9,<3.13"  # Compatible python versions must be declared here
 pandas = "1.5.3"
 psutil = "^5.8.0"
 glob2 = "^0.7"
