### 2025-04-24 #7817 @rmndrs89
Hi @sappelhoff and @Trybnetic

I have had a quick look at the repo and the draft of the paper. It certainly looks interesting to me, but I am also sharing the same concerns as @sappelhoff .

For example, there is already [`actipy`](https://github.com/OxWearables/actipy) that has similar functionality, i.e. processing accelerometer data ([resampling](https://github.com/OxWearables/actipy/blob/879e0e72dcca8069dd893b3f20ef66cb5f7884fb/src/actipy/processing.py#L94), [filtering](https://github.com/OxWearables/actipy/blob/879e0e72dcca8069dd893b3f20ef66cb5f7884fb/src/actipy/processing.py#L173), [detecting non-wear](https://github.com/OxWearables/actipy/blob/879e0e72dcca8069dd893b3f20ef66cb5f7884fb/src/actipy/processing.py#L556)). Would it no be easier to contribute to existing packages? Would that not be more in the spirit of open-source packaging?

If no other reviewers are found before mid June, I can do it, but I have limited time before that because of other deadlines and annual leave.
