### 2025-04-24 #7817 @Trybnetic
Hei @rmndrs89,

I can absolutely follow you concerns.

I am following `actipy` since it was made public, because I also initially had the impression that both packages might have a similar scope. However, the impression that I got from `actipy` the last years is, that it focuses more on the data loading and cleaning than on the further processing of the data and that there is no thrive to change this. E.g., they  have one non wear time method implemented, but it is not further described which method this is and as far as I can see it is the only method beyond reading, resampling and filtering which they have implemented.

Simultaneously, they implemented a very good reading and autocalibration functionality which we in fact started using and stopped implementing again for `paat`., because it seemed unnecessary to have two implementations.

Personally, I am very open to joining forces of any kind if it makes sense and I am very happy about their contribution and if they in the future plan to shift/extend the focus of their package, we are also more than open to join forces. However, as I perceive the situation currently the ideas/scopes of the packages are different and I am therefore unsure whether this makes sense currently.
