### 2025-04-24 #7817 @AKuederle
To add to this: The successor to `gaitpy`, [scikit digital health](https://scikit-digital-health.readthedocs.io/en/latest/) also seems to have a quite a bit of overlap in functionality.

This does not mean that PAAT "should not exist", just that the paper (and maybe the README) should make it clear how it differentiates and where the authors saw a need beyond the available solutions.
