### 2025-04-24 #7817 @sappelhoff
@Trybnetic please keep us up to date once you have implemented these updates.

@AKuederle thank you for your input, would you like to also further review this package?

@rmndrs89 thank you for your input, too. And also for your offer. I hope to find peer reviewers sooner than mid June, but will come back to you in case we do not find anyone.
