### 2025-04-27 #7817 @mprib
@Trybnetic,

As I get up to speed for this review, are there open source hip worn accelerometry datasets that I could use on my end for independent testing?

Thank you,

Mac
