### 2025-04-27 #7817 @Trybnetic
Hei @mprib,

I am not aware of any open source datasets including raw hip acceleration data. This would in fact have been a delight for the method development. I am only aware of a few data sets with activity count data, but they are sadly not relevant for this package.

We have included [one day of data](https://github.com/Trybnetic/paat/blob/main/docs/source/example_notebooks/data/example_day_calibrated.csv.tar.gz) in the documentation, which is also used in the [Analyze data with PAAT](https://paat.readthedocs.io/en/latest/example_notebooks/analyze_data.html#Analyze-data-with-PAAT) notebook. There is also [limited data from one of our studies](https://doi.org/10.18710/VP5DMU) available. However, in both cases we did not provide gt3x files as they in these cases contained personal information. For the IO functionality, there is also some [test data](https://github.com/Trybnetic/paat/blob/main/docs/source/example_notebooks/data/10min_recording.gt3x) provided in the repository.

I hope this might at least partially help you
