### 2025-04-28 #7817 @mprib
Thank you for the links, @Trybnetic. This is extremely helpful for quickly getting oriented to the project. I completely understand (and sympathize) with the challenges of finding good open source datasets.
