### 2025-04-28 #7817 @sappelhoff
Thanks for the note @marcoghislieri.

👋  @LukasAdamowicz, @matt002

would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
