+++ paat/sleep.py
@@ -13,7 +13,8 @@ os.environ['TF_CPP_MIN_LOG_LEVEL'] = '3'
 import pandas as pd
 import numpy as np
 import tensorflow as tf
-from tensorflow.keras import models
+# from tensorflow.keras import models
+import keras
 
 # Hide GPU from visible devices
 tf.config.set_visible_devices([], 'GPU')
@@ -65,10 +66,15 @@ def detect_time_in_bed_weitz2024(data, sample_freq, resampled_frequency="1min", 
 
     # Load model if not specified
     if not model:
-        model_path = os.path.join(os.path.pardir, os.path.dirname(__file__), 'models', 'TIB_model.h5')
-        model = models.load_model(model_path)
+        # model_path = os.path.join(os.path.pardir, os.path.dirname(__file__), 'models', 'TIB_model.h5')
+        # model = models.load_model(model_path)
 
-    predictions = (model.predict(X[np.newaxis], verbose=0).squeeze() >= .5)
+        model_path = os.path.join(os.path.pardir, os.path.dirname(__file__), 'models', 'TIB_model.pb')
+        model= keras.layers.TFSMLayer(model_path, call_endpoint='serving_default')
+
+    # predictions = (model.predict(X[np.newaxis], verbose=0).squeeze() >= .5)
+    # predictions = (model(X[np.newaxis]).squeeze() >= .5)
+    predictions = (model(X[np.newaxis])["output_0"].numpy().squeeze() >= .5)
 
     seconds = pd.Timedelta(resampled_frequency).seconds
     predictions = np.repeat(predictions, seconds * sample_freq)
