### 2025-04-29 bcbf8cd
enables python 3.12 testing
