+++ .github/workflows/python-test.yml
@@ -18,7 +18,7 @@ jobs:
           #windows-latest, 
           #macOS-latest
           ]
-        python-version: ['3.10', '3.11']
+        python-version: ['3.10', '3.11', '3.12']
         poetry-version: [1.7.1]
       fail-fast: false
     runs-on: ${{ matrix.os }}
