### 2025-04-29 2d21d28
enables using newer resampy versions
