+++ pyproject.toml
@@ -30,7 +30,7 @@ classifiers = [
 python = ">3.9,<3.13"  # Compatible python versions must be declared here
 psutil = "^5.8.0"
 glob2 = "^0.7"
-resampy = "^0.2.2"
+resampy = "^0"
 joblib = "^1.0.1"
 bitstring = "^3.1.7"
 agcounts = "^0.2"
