### 2025-05-05 f2bf96b
Ports code and models to keras3 (#45)
* converts models to new .pb file format

* ports code to tensorflow>2.15 and keras>=3.0

* adds python 3.12 testing

* adds slow test to compare outputs with last keras2 version

* bumps version
