+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "paat"
-version = "1.0.0b8"
+version = "1.0.0b9"
 description = "A comprehensive toolbox to analyse and model raw physical activity data"
 
 license = "MIT"
@@ -27,19 +27,17 @@ classifiers = [
 ]
 
 [tool.poetry.dependencies]
-python = ">3.9,<3.12"  # Compatible python versions must be declared here
-pandas = "1.5.3"
+python = ">3.9,<3.13"  # Compatible python versions must be declared here
 psutil = "^5.8.0"
 glob2 = "^0.7"
-resampy = "^0.2.2"
+resampy = "^0"
 joblib = "^1.0.1"
-numpy = "^1.23"
 bitstring = "^3.1.7"
-tensorflow = ">=2.11.0,<2.16"
 agcounts = "^0.2"
 toml = "^0.10.2"
 tables = "^3.7.0"
 pygt3x = "^0.7.1"
+tensorflow = "^2.16.1"
 
 [tool.poetry.extras]
 docs = ["sphinx", "sphinx_rtd_theme", "numpydoc", "easydev", "nbsphinx", "docutils"]
