### 2025-05-08 7d9a17d
extends discussion of other accelerometry packages
