+++ paper/paper.bib
@@ -138,16 +138,26 @@ Section: Journal for the Measurement of Physical Behaviour},
 	file = {Snapshot:/home/msw/Zotero/storage/HXYIKBTP/14514694.html:text/html},
 }
 
-@article{weitz_automatic_2024,
-	title = {Automatic time in bed detection from hip-worn accelerometers for large epidemiological studies: {The} {Tromso} {Study}},
-	url = {http://medrxiv.org/content/early/2025/02/19/2022.03.07.22270992.abstract},
-	doi = {10.1101/2022.03.07.22270992},
-	abstract = {Accelerometers are frequently used to assess physical activity in large epidemiological studies. They can monitor movement patterns and cycles over several days under free-living conditions and are usually either worn on the wrist or the hip. While wrist-worn accelerometers have been frequently used to additionally assess sleep and time in bed behavior, hip-worn accelerometers have been widely neglected for this task due to their primary focus on physical activity. Here, we present a new method with the objective to identify the time in bed to enable further analysis options for large-scale studies using hip-placement like time in bed or sedentary time analyses. We introduced new and accelerometer specific data augmentation methods, such as mimicking a wrongly worn accelerometer, additional noise, and random croping, to improve training and generalization performance. Subsequently, we trained a neural network model on a sample from the population-based Tromsø Study and evaluated it on two additional datasets. Our algorithm achieved an accuracy of 94\% on the training data, 92\% on unseen data from the same population and comparable results to consumer-wearable data obtained from a demographically different population. Generalization performance was overall good, however, we found that on a few particular days or participants, the trained model fundamentally over- or underestimated time in bed (e.g., predicted all or nothing as time in bed). Despite these limitations, we anticipate our approach to be a starting point for more sophisticated methods to identify time in bed or at some point even sleep from hip-worn acceleration signals. This can enable the re-use of already collected data, for example, for longitudinal analyses where sleep-related research questions only recently got into focus or sedentary time needs to be estimated in 24h wear protocols.},
-	journal = {medRxiv},
-	author = {Weitz, Marc and Syed, Shaheen and Hopstock, Laila Arnesdatter and Morseth, Bente and Henriksen, Andre and Horsch, Alexander},
-	month = jan,
+
+@article{weitz_automatic_2025,
+	title = {Automatic time in bed detection from hip-worn accelerometers for large epidemiological studies: {The} {Tromsø} {Study}},
+	volume = {20},
+	issn = {1932-6203},
+	shorttitle = {Automatic time in bed detection from hip-worn accelerometers for large epidemiological studies},
+	url = {https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0321558},
+	doi = {10.1371/journal.pone.0321558},
[…]
