### 2025-05-08 2d914b7
enables windows and macos testing
