+++ .github/workflows/python-test.yml
@@ -15,8 +15,8 @@ jobs:
       matrix:
         os: [
           ubuntu-latest, 
-          #windows-latest, 
-          #macOS-latest
+          windows-latest, 
+          macOS-latest
           ]
         python-version: ['3.10', '3.11', '3.12']
         poetry-version: [2.1.3]
