### 2025-05-09 d61eb0e
fixes equation formatting
