### 2025-05-09 d61b54d
Update docs (#46)
* updates documentation

* updates tib reference

* updates testing workflow and includes Windows and macos testing
