### 2025-05-09 0ff4ed5
Update readme (#47)
* updates readme and documentation home page
