### 2025-05-09 #8136 @Trybnetic
@mprib, @LukasAdamowicz & @AKuederle: I have now updated the manuscript, readme and documentation and also updated the code to enable using the package on Python 3.12. I am looking forward to your review.
