### 2025-06-06 #8136 @LukasAdamowicz
Comments from initial/start of review:

- Column name requirement is not specified in the README (there is a statement about time as index to a pandas DataFrame, but no other specification). Additionally, it does not specify that time should be human timestamps (ie not epoch timestamps in s/ms/ns etc)
- Handling of various timestamps seems off somehow for the `detect_time_in_bed_weitz2024` method. I can't get it to return an array of the same size, even trying on a tiny fraction of data (100 seconds), from any of 3 devices I have some data for.
- sleep.py line 56: `X = data.reset_index()[["Y", "X", "Z"]].values.copy()` should definitely have a comment explaining why you are re-ordering axes from the accelerometer.
- Along these lines, you should have specification of which axes correspond to which directions. If you want anyone to contribute to this, either axes identification needs to happen automatically, or it needs to be very clearly and carefully documented where it is easily accessible. Not all devices are going to use the same axes orientations within the device housing. If it is axis-agnostic (ie using vector magnitude) this should be also clarified
