### 2025-06-11 Trybnetic/paat#49 @Trybnetic
Column name requirement is not specified
Column name requirement is not specified in the README (there is a statement about time as index to a pandas DataFrame, but no other specification). Additionally, it does not specify that time should be human timestamps (ie not epoch timestamps in s/ms/ns etc)
