### 2025-06-11 Trybnetic/paat#50 @Trybnetic
Handling of various timestamps seems off
Handling of various timestamps seems off somehow for the detect_time_in_bed_weitz2024 method. I can't get it to return an array of the same size, even trying on a tiny fraction of data (100 seconds), from any of 3 devices (including actigraph, but csv format) I have some data for.
