### 2025-06-11 Trybnetic/paat#51 @Trybnetic
comment explaining why you are re-ordering axes
sleep.py line 56: X = data.reset_index()[["Y", "X", "Z"]].values.copy() should definitely have a comment explaining why you are re-ordering axes from the accelerometer.
