### 2025-06-11 Trybnetic/paat#52 @Trybnetic
specification of which axes correspond to which directions
Along these lines, you should have specification of which axes correspond to which directions. If you want anyone to contribute to this, either axes identification needs to happen automatically, or it needs to be very clearly and carefully documented where it is easily accessible. Not all devices are going to use the same axes orientations within the device housing. If it is axis-agnostic (ie using vector magnitude) this should be also clarified
