### 2025-06-11 18624c2
adds comment on why axes are reordered (fixes #51)
