### 2025-06-11 ce68236
adds info that YXZ acceleration columns are required (fixes #49) and which axis refers to which axes of the human body (fixes #52)
