### 2025-06-14 #8136 @Trybnetic
@sappelhoff @LukasAdamowicz I have addressed now issues addressed the issues [#49](https://github.com/Trybnetic/paat/issues/49), [#51](https://github.com/Trybnetic/paat/issues/51), & [#52](https://github.com/Trybnetic/paat/issues/52) in the paper branch. With [#50](https://github.com/Trybnetic/paat/issues/50), I am not sure what the problem here is. @LukasAdamowicz are you able to share a minimal example of what does not work?
