### 2025-06-14 22c3768
adds cutting predictions if no full minutes are provided
