### 2025-06-14 Trybnetic/paat#50 @Trybnetic
@LukasAdamowicz I am still trying to understand the problem here. When looking at the code I realized that a problem can be if the provided data does not contain full minutes. I now added that in this case the predictions get matched to the provided data. However, I am not sure whether this is your issue, so I would be glad if you could provide a minimal example to showcase what is not working
