### 2025-06-16 #8136 @LukasAdamowicz
> [@sappelhoff](https://github.com/sappelhoff) [@LukasAdamowicz](https://github.com/LukasAdamowicz) I have addressed now issues addressed the issues [#49](https://github.com/Trybnetic/paat/issues/49), [#51](https://github.com/Trybnetic/paat/issues/51), & [#52](https://github.com/Trybnetic/paat/issues/52) in the paper branch. With [#50](https://github.com/Trybnetic/paat/issues/50), I am not sure what the problem here is. [@LukasAdamowicz](https://github.com/LukasAdamowicz) are you able to share a minimal example of what does not work?

```python
n_min = 3
fs = 64

t0 = pd.to_datetime("2025-06-10 13:04:57")
td = pd.to_timedelta(np.arange(0, n_min * 60, 1/fs), unit='s')

time = t0 + td

acc = np.random.default_rng().normal(size=(time.size, 3))
acc[:, 1] += 1

df = pd.DataFrame(data=acc, columns=['X', 'Y', 'Z'])
df.index = time

df.head(3)

df.loc[:, "Non Wear Time"] = paat.detect_non_wear_time_hees2011(df, 64)

df.loc[:, "Time in Bed"] = paat.detect_time_in_bed_weitz2024(df, 64)
```
