### 2025-07-11 92df262
Clarifies the value of hip-worn accelermeters (#53) and amends paper by recent research
