+++ paper/paper.bib
@@ -388,3 +388,42 @@ Publisher: JMIR Publications Inc., Toronto, Canada},
 	pages = {e36762},
 	file = {Full Text:/home/msw/Zotero/storage/KZ5N98J6/Adamowicz et al. - 2022 - SciKit Digital Health Python Package for Streamli.pdf:application/pdf;Snapshot:/home/msw/Zotero/storage/EFXBESF5/e36762.html:text/html},
 }
+
+@article{swartz_estimation_2000,
+	title = {Estimation of energy expenditure using {CSA} accelerometers at hip and wrist sites},
+	volume = {32},
+	issn = {0195-9131},
+	url = {https://journals.lww.com/acsm-msse/fulltext/2000/09001/estimation_of_energy_expenditure_using_csa.3.aspx},
+	abstract = {SWARTZ, A. M., S. J. STRATH, D. R. BASSETT, JR., W. L. O’BRIEN, G. A. KING, and B. E. AINSWORTH. Estimation of energy expenditure using CSA accelerometers at hip and wrist sites. Med. Sci. Sports Exerc., Vol. 32, No. 9, Suppl., pp. S450–S456, 2000.
+          Purpose 
+          This study was designed to establish prediction models that relate hip and wrist accelerometer data to energy expenditure (EE) in field and laboratory settings. We also sought to determine whether the addition of a wrist accelerometer would significantly improve the prediction of EE (METs), compared with a model that used a hip accelerometer alone.
+          Methods 
+          Seventy participants completed one to six activities within the categories of yardwork, housework, family care, occupation, recreation, and conditioning, for a total of 5 to12 participants tested per activity. EE was measured using the Cosmed K4b2 portable metabolic system. Simultaneously, two Computer Science and Applications, Inc. (CSA) accelerometers (model 7164), one worn on the wrist and one worn on the hip, recorded body movement. Correlations between EE measured by the Cosmed and the counts recorded by the CSA accelerometers were calculated, and regression equations were developed to predict EE from the CSA data.
+          Results 
+          The wrist, hip, and combined hip and wrist regression equations accounted for 3.3\%, 31.7\%, and 34.3\% of the variation in EE, respectively. The addition of the wrist accelerometer data to the hip accelerometer data to form a bivariate regression equation, although statistically significant (P = 0.002), resulted in only a minor improvement in prediction of EE. Cut points for 3 METs (574 hip counts), 6 METs (4945 hip counts), and 9 METs (9317 hip counts) were also established.
+          Conclusion 
+          The small amount of additional accuracy gained from the wrist accelerometer is offset by the extra time required to analyze the data and the cost of the accelerometer.},
+	language = {en-US},
+	number = {9},
+	urldate = {2024-04-26},
+	journal = {Medicine \& Science in Sports \& Exercise},
+	author = {Swartz, Ann M. and Strath, Scott J. and Bassett, David R. and O’brien, William L. and King, George A. and Ainsworth, Barbara E.},
+	month = sep,
+	year = {2000},
+	keywords = {thesis},
+	pages = {S450},
+	file = {Snapshot:/home/msw/Zotero/storage/WFJAMF2T/estimation_of_energy_expenditure_using_csa.3.html:text/html},
+}
+
+@article {weitz_comparison_2025,
+	author = {Weitz, Marc and Morseth, Bente and Hopstock, Laila A. and Brage, S{\o}ren and Horsch, Alexander},
+	title = {Comparison of sedentary time calculated from count- and raw accelerometer data: The Troms{\o} Study},
+	elocation-id = {2025.06.02.25328840},
+	year = {2025},
+	doi = {10.1101/2025.06.02.25328840},
+	publisher = {Cold Spring Harbor Laboratory Press},
[…]
