+++ paper/paper.md
@@ -46,14 +46,14 @@ noncommunicable diseases such as cancer, cardiovascular diseases, or diabetes as
 or anxiety. Various methods exist to measure physical activity. One of these methods use accelerometers to estimate
 physical activity. Accelerometers are small body-worn sensors which measure acceleration over time. They have become a
 popular assessment tool in research and public health as they provide a reasonably cheap but still more objective
-alternative to surveys while simultaneously keeping the researcher and participant burden low. Accelerometers measure the raw acceleration in ms$^{−2}$ often also expressed as multiples of Earth's gravitation ($1g =  9.80665ms^{−2}$). However, due to historic limitations of on-device storage, the raw acceleration has often been processed to summary metrics like activity counts [@neishabouri_quantification_2022]. However, over the last decade the raw acceleration itself has also gotten into the focus of method development [@van_hees_challenges_2016]. A fundamental limitation of accelerometry is that many methods are only applicable to a certain wear location and demographic group and often do not generalize beyond that. Therefore, separate sets of methods have been developed for different populations and study protocols.
+alternative to surveys while simultaneously keeping the researcher and participant burden low. Accelerometers measure the raw acceleration in ms$^{−2}$ often also expressed as multiples of Earth's gravitation ($1g =  9.80665ms^{−2}$). However, due to historic limitations of on-device storage, the raw acceleration has often been processed to summary metrics like activity counts [@neishabouri_quantification_2022]. Over the last decade the raw acceleration itself has also gotten into the focus of method development [@van_hees_challenges_2016]. A fundamental limitation of accelerometry is that many methods are only applicable to a certain wear location and demographic group and often do not generalize beyond that. All wear locations come with advantages and disadvantages. Hip-based accelerometer assessment shows higher correlation with physical activity energy expenditure due to its proximity to the body's center of mass [@swartz_estimation_2000] while wrist-worn accelerometers are perceived as less obtrusive and thus has been argued to increase study protocol adherence [@troiano_evolution_2014]. Unfortunately, separate sets of methods have been developed for different populations, study protocols, and wear locations with most methods and analysis tools focusing on wrist-worn accelerometers.
 
 <!-- Accelerometry packages -->
-Today, a plethora of accelerometer packages exist each fulfilling different purposes. The most popular package is GGIR [@migueles_ggir_2019] which provides a broad set of well validated raw data methods mainly focusing on wrist-based accelerometry in R. @hammad_pyactigraphy_2021 implemented *pyactigraphy* to analyze accelerometer activity count and light exposure data. *Actipy* provides various file reading and autocalibration functions, but only provide a limited set of methods for the further data analysis [@chan_actipy_2024]. *SciKit Digital Health (SKDH)* provides a variety of algorithms for deriving clinical features of gait, sit to stand, physical activity, and sleep [@adamowicz_scikit_2022]. *Paat* is more specific in its focus on 
+Today, a plethora of accelerometer packages exist each fulfilling different purposes. The most popular package is GGIR [@migueles_ggir_2019] which provides a broad set of well validated raw data methods mainly focusing on wrist-based accelerometry in R. @hammad_pyactigraphy_2021 implemented *pyactigraphy* to analyze accelerometer activity count and light exposure data. 
 
 ![Visualization of the results obtained from *paat*. (A) The package can be used to load and process the raw data (upper row). The loaded data can then be annotated by a variety of methods. The implemented non-wear time and time in bed algorithm exploit raw acceleration data directly. To estimate physical activity, the raw data is reduced to the ENMO of the signal (lower row). Alternatively, also other metrics like MAD can be estimated and used for further processing. (B) Aggregated daily or average (Ø) results can be obtained and then be used for further analyzes. \label{fig:processing}](img/paper_fig1.png) 
 
[…]
