### 2025-07-11 #8136 @Trybnetic
@LukasAdamowicz have you tried your example with the latest version of the paper branch? For me your example produces the expected behavior:

``` bash
>>> df
                                   X         Y         Z  Non Wear Time  Time in Bed
2025-06-10 13:04:57.000000  0.211491  1.213597 -0.900728          False        False
2025-06-10 13:04:57.015625  1.135398  0.895683 -0.018674          False        False
2025-06-10 13:04:57.031250  1.231842  3.042498 -1.237272          False        False
2025-06-10 13:04:57.046875 -0.660009 -0.550419  1.298916          False        False
2025-06-10 13:04:57.062500  0.919274 -1.356896  0.033287          False        False
...                              ...       ...       ...            ...          ...
2025-06-10 13:07:56.921875 -0.181678  1.194814  0.306432          False        False
2025-06-10 13:07:56.937500 -0.625516  1.042836  0.750449          False        False
2025-06-10 13:07:56.953125 -0.396959  0.932563  0.332425          False        False
2025-06-10 13:07:56.968750  1.085966  1.196715 -0.789999          False        False
2025-06-10 13:07:56.984375  2.139442  1.610082  0.210717          False        False

[11520 rows x 5 columns]

```
