### 2025-07-11 #8136 @sappelhoff
Thanks for the update @Trybnetic.

And yes, @mprib @LukasAdamowicz would be great if you could raise remaining issues if any -- and complete your review at your next convenience 🙂 Please let us know about the tentative timetable.

@mprib I assume that your original plan from https://github.com/openjournals/joss-reviews/issues/8136#issuecomment-2972781691 got off track?
