### 2025-07-15 67a543a
adds the MWE suggested by @LukasAdamowicz
