### 2025-07-15 2b659a5
activates tests for mwe expample
