### 2025-07-15 #8136 @LukasAdamowicz
<summary>Error</summary>

Cell In[6], line 1
----> 1 df.loc[:, "Time in Bed"] = paat.detect_time_in_bed_weitz2024(df, 64)

File ~/opt/miniconda3/envs/paat/lib/python3.12/site-packages/pandas/core/indexing.py:911, in _LocationIndexer.__setitem__(self, key, value)
    908 self._has_valid_setitem_indexer(key)
    910 iloc = self if self.name == "iloc" else self.obj.iloc
--> 911 iloc._setitem_with_indexer(indexer, value, self.name)

File ~/opt/miniconda3/envs/paat/lib/python3.12/site-packages/pandas/core/indexing.py:1857, in _iLocIndexer._setitem_with_indexer(self, indexer, value, name)
   1854 # add a new item with the dtype setup
   1855 if com.is_null_slice(indexer[0]):
   1856     # We are setting an entire column
-> 1857     self.obj[key] = value
   1858     return
   1859 elif is_array_like(value):
   1860     # GH#42099

File ~/opt/miniconda3/envs/paat/lib/python3.12/site-packages/pandas/core/frame.py:4316, in DataFrame.__setitem__(self, key, value)
   4313     self._setitem_array([key], value)
   4314 else:
   4315     # set column
-> 4316     self._set_item(key, value)

File ~/opt/miniconda3/envs/paat/lib/python3.12/site-packages/pandas/core/frame.py:4529, in DataFrame._set_item(self, key, value)
   4519 def _set_item(self, key, value) -> None:
   4520     """
   4521     Add series to DataFrame in specified column.
   4522
   (...)   4527     ensure homogeneity.
   4528     """
-> 4529     value, refs = self._sanitize_column(value)
   4531     if (
   4532         key in self.columns
   4533         and value.ndim == 1
   4534         and not isinstance(value.dtype, ExtensionDtype)
   4535     ):
   4536         # broadcast across multiple columns if necessary
   4537         if not self.columns.is_unique or isinstance(self.columns, MultiIndex):

File ~/opt/miniconda3/envs/paat/lib/python3.12/site-packages/pandas/core/frame.py:5273, in DataFrame._sanitize_column(self, value)
   5270     return _reindex_for_setitem(value, self.index)
   5272 if is_list_like(value):
-> 5273     com.require_length_match(value, self.index)
   5274 arr = sanitize_array(value, self.index, copy=True, allow_2d=True)
   5275 if (
   5276     isinstance(value, Index)
   5277     and value.dtype == "object"
   (...)   5280     # TODO: Remove kludge in sanitize_array for string mode when enforcing
   5281     # this deprecation

File ~/opt/miniconda3/envs/paat/lib/python3.12/site-packages/pandas/core/common.py:573, in require_length_match(data, index)
    569 """
    570 Check the length of data matches the length of the index.
    571 """
    572 if len(data) != len(index):
--> 573     raise ValueError(
    574         "Length of values "
    575         f"({len(data)}) "
    576         "does not match length of index "
    577         f"({len(index)})"
    578     )

ValueError: Length of values (15360) does not match length of index (11520)

Output of paat.sysinfo (not using poetry, just imported where I'm running the MWE):
```
paat.sysinfo()

PAAT Information
=================

General Information
-------------------
Python version: 3.12.0
PAAT version: 1.0.0b9

Operating System
----------------
OS: Darwin arm64
Kernel: 24.3.0

Dependencies
------------
agcounts: 0.2.6
bitstring: 3.1.9
glob2: 0.7
joblib: 1.5.1
psutil: 5.9.8
pygt3x: 0.7.1
resampy: 0.4.3
tables: 3.10.2
tensorflow: 2.19.0
toml: 0.10.2
```
