### 2025-07-15 #8136 @Trybnetic
I can reproduce your error with the version in the main branch, but it seems fixed to me in the paper branch (at least there it works without any error for me and in the Github action). My current theory is that it might not have properly changed paat to the version of the paper branch as it there still has the same version tag as in the main branch. I have bumped the version of the paper branch to `1.0.0b10` and hope that you do not longer get the error under this version.
