### 2025-07-15 #8136 @LukasAdamowicz
> I can reproduce your error with the version in the main branch, but it seems fixed to me in the paper branch (at least there it works without any error for me and in the Github action). My current theory is that it might not have properly changed paat to the version of the paper branch as it there still had the same version tag as in the main branch. I have bumped the version of the paper branch to `1.0.0b10` now and hope that you do not longer get the error under this version. Please let me know if you still get the error for `v1.0.0b10`

I'll try re-installing again and see what happens
