### 2025-07-15 #8136 @sappelhoff
@Trybnetic  please work on your points from https://github.com/openjournals/joss-reviews/issues/8136#issuecomment-3075417571 and let me know once you are done.

I will then continue with my list.
