### 2025-07-16 e8cc1ea
bumps version to v1.0.0
