+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "paat"
-version = "1.0.0b10"
+version = "1.0.0"
 description = "A comprehensive toolbox to analyse and model raw physical activity data"
 
 license = "MIT"
