### 2025-07-16 f0a0f54
sets version to v1.0.0b10 to test that zenodo export is still working after changes to zenodo.json
