### 2025-07-16 f9863d6
Adds JOSS paper (#54)
* adds JOSS paper manuscript

* corrects workflow versions

* updates zenodo title to paper title and adds orcid ids

* bumps version to v1.0.0b10
