+++ paper/paper.md
@@ -1,0 +1,76 @@
+---
+title: 'PAAT: Physical Activity Analysis Toolbox for the analysis of hip-worn raw accelerometer data in Python'
+tags:
+  - physical-activity
+  - accelerometers
+  - ActiGraph
+  - hip
+  - methods
+authors:
+  - name: Marc Weitz
+    orcid: 0009-0006-8225-833X
+    corresponding: true # (This is how to denote the corresponding author)
+    equal-contrib: false
+    affiliation: "1" # (Multiple affiliations must be quoted)
+  - name: Shaheen Syed
+    orcid: 0000-0001-5462-874X
+    equal-contrib: false # (This is how you can denote equal contributions between multiple authors)
+    affiliation: "1"
+  - name: Alexander Horsch
+    orcid: 0000-0001-7745-0139
+    affiliation: "1"
+affiliations:
+ - name: Department of Computer Science, UiT The Arctic University of Norway, Tromsø, Norway
+   index: 1
+date: 01 May 2024
+bibliography: paper.bib
+
+---
+
+# Summary
+
+The Physical Activity Analysis Toolbox (*paat*) is a versatile Python package designed to analyze physical activity
+data. It is used in research and health-related fields to process raw acceleration data collected from the hip. It
+supports importing, cleaning, and preprocessing raw data, and includes algorithms to classify time in bed and non wear
+time. Furthermore, it supports estimating various physical activity levels such as moderate-to-vigorous physical
+activity or sedentary behavior, with customizable thresholds from multiple metrics. These estimates can be aggregated
+and used for further statistical analysis or be used directly for more sophisticated physical activity pattern analysis.
+Additionally, *paat* is extensible, allowing users to add custom algorithms or modules, and it integrates well with
+other data analysis tools within the Python ecosystem. 
+
+# Statement of need
+
+<!-- Measurement of physical activity -->
+Physical activity is one of the strongest predictors of overall health. Its absence has been linked to various
+noncommunicable diseases such as cancer, cardiovascular diseases, or diabetes as well as mental diseases like depression
+or anxiety. Various methods exist to measure physical activity. One of these methods use accelerometers to estimate
+physical activity. Accelerometers are small body-worn sensors which measure acceleration over time. They have become a
+popular assessment tool in research and public health as they provide a reasonably cheap but still more objective
+alternative to surveys while simultaneously keeping the researcher and participant burden low. Accelerometers measure the raw acceleration in ms$^{−2}$ often also expressed as multiples of Earth's gravitation ($1g =  9.80665ms^{−2}$). However, due to historic limitations of on-device storage, the raw acceleration has often been processed to summary metrics like activity counts [@neishabouri_quantification_2022]. Over the last decade the raw acceleration itself has also gotten into the focus of method development [@van_hees_challenges_2016]. A fundamental limitation of accelerometry is that many methods are only applicable to a certain wear location and demographic group and often do not generalize beyond that. All wear locations come with advantages and disadvantages. Hip-based accelerometer assessment shows higher correlation with physical activity energy expenditure due to its proximity to the body's center of mass [@swartz_estimation_2000] while wrist-worn accelerometers are perceived as less obtrusive and thus has been argued to increase study protocol adherence [@troiano_evolution_2014]. Unfortunately, separate sets of methods have been developed for different populations, study protocols, and wear locations with most methods and analysis tools focusing on wrist-worn accelerometers.
+
+<!-- Accelerometry packages -->
+Today, a plethora of accelerometer packages exist each fulfilling different purposes. The most popular package is GGIR [@migueles_ggir_2019] which provides a broad set of well validated raw data methods mainly focusing on wrist-based accelerometry in R. @hammad_pyactigraphy_2021 implemented *pyactigraphy* to analyze accelerometer activity count and light exposure data. 
+
[…]
