+++ paper/paper.bib
@@ -1,0 +1,429 @@
+@article{syed_evaluating_2020,
+	title = {Evaluating the performance of raw and epoch non-wear algorithms using multiple accelerometers and electrocardiogram recordings},
+	volume = {10},
+	copyright = {2020 The Author(s)},
+	issn = {2045-2322},
+	url = {https://www.nature.com/articles/s41598-020-62821-2},
+	doi = {10.1038/s41598-020-62821-2},
+	abstract = {Accurate detection of accelerometer non-wear time is crucial for calculating physical activity summary statistics. In this study, we evaluated three epoch-based non-wear algorithms (Hecht, Troiano, and Choi) and one raw-based algorithm (Hees). In addition, we performed a sensitivity analysis to provide insight into the relationship between the algorithms’ hyperparameters and classification performance, as well as to generate tuned hyperparameter values to better detect episodes of wear and non-wear time. We used machine learning to construct a gold-standard dataset by combining two accelerometers and electrocardiogram recordings. The Hecht and Troiano algorithms achieved poor classification performance, while Choi exhibited moderate performance. Meanwhile, Hees outperformed all epoch-based algorithms. The sensitivity analysis and hyperparameter tuning revealed that all algorithms were able to achieve increased classification performance by employing larger intervals and windows, while more stringently defining artificial movement. These classification gains were associated with the ability to lower the false positives (type I error) and do not necessarily indicate a more accurate detection of the total non-wear time. Moreover, our results indicate that with tuned hyperparameters, epoch-based non-wear algorithms are able to perform just as well as raw-based non-wear algorithms with respect to their ability to correctly detect true wear and non-wear episodes.},
+	language = {en},
+	number = {1},
+	urldate = {2020-12-03},
+	journal = {Scientific Reports},
+	author = {Syed, Shaheen and Morseth, Bente and Hopstock, Laila A. and Horsch, Alexander},
+	month = apr,
+	year = {2020},
+	note = {Number: 1
+Publisher: Nature Publishing Group},
+	keywords = {paat-paper, paper1, paper3},
+	pages = {1--18},
+	file = {Full Text PDF:/home/msw/Zotero/storage/YKGWM48N/Syed et al. - 2020 - Evaluating the performance of raw and epoch non-we.pdf:application/pdf;Snapshot:/home/msw/Zotero/storage/KPCCT6WI/s41598-020-62821-2.html:text/html;Snapshot:/home/msw/Zotero/storage/7MKIWD4E/s41598-020-62821-2.html:text/html},
+}
+
+@article{van_hees_estimation_2011,
+	title = {Estimation of {Daily} {Energy} {Expenditure} in {Pregnant} and {Non}-{Pregnant} {Women} {Using} a {Wrist}-{Worn} {Tri}-{Axial} {Accelerometer}},
+	volume = {6},
+	issn = {1932-6203},
+	url = {https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0022922},
+	doi = {10.1371/journal.pone.0022922},
[…]
